Updated plugins

+++ popup-builder/com/classes/components/Menu.php
@@ -115,7 +115,7 @@ class SGPBMenu
 	 * @return false|int
 	 */
 	private function inMultiDimArray($item , $array){
-		return preg_match('/"'.preg_quote($item, '/').'"/i' , json_encode($array));
+		return preg_match('/"'.preg_quote($item, '/').'"/i' , wp_json_encode($array));
 	}
 
 	/**
@@ -202,7 +202,7 @@ class SGPBMenu
     { ?>
         <div class="description  description-wide">
             <label for="edit-menu-item-pb-<?php echo esc_attr($item->ID); ?>">
-                <?php esc_html_e('Select a Popup', SG_POPUP_TEXT_DOMAIN); ?><br/>
+                <?php esc_html_e('Select a Popup', 'popup-builder'); ?><br/>
                 <select class="widefat" name="menu-item-pb[<?php echo esc_attr($item->ID); ?>][popup]"
                         id="edit-menu-item-pb-<?php echo esc_attr($item->ID); ?>">
                     <option value=""></option>
@@ -212,7 +212,7 @@ class SGPBMenu
                         </option>
                     <?php endforeach; ?>
                 </select>
-                <span class="description"><?php esc_html_e('Open a popup once this item is clicked.', SG_POPUP_TEXT_DOMAIN); ?></span>
+                <span class="description"><?php esc_html_e('Open a popup once this item is clicked.', 'popup-builder'); ?></span>
             </label>
         </div>
         <?php
@@ -226,10 +226,13 @@ class SGPBMenu
      */
     public function save($menu_id, $item_id)
     {
-        delete_post_meta($item_id, '_menu_sgpb_popup_id');
-        if (isset($_POST['menu-item-pb'][$item_id]['popup'])) {
-            $popupId = (int)sanitize_text_field($_POST['menu-item-pb'][$item_id]['popup']);
-            update_post_meta($item_id, '_menu_sgpb_popup_id', $popupId);
-        }
+        if ( current_user_can( 'edit_theme_options' ) ) {
+			check_admin_referer( 'update-nav_menu', 'update-nav-menu-nonce' );
+			delete_post_meta($item_id, '_menu_sgpb_popup_id');
+			if (isset($_POST['menu-item-pb'][$item_id]['popup'])) {
+				$popupId = (int)sanitize_text_field($_POST['menu-item-pb'][$item_id]['popup']);
+				update_post_meta($item_id, '_menu_sgpb_popup_id', $popupId);
+			}
+		}
     }
 }
